fix: swallowed exceptions from improper exports
Adapters were causing multiple instances of OAuthExceptions to be created, one for each instance.
Instead of going down the bundler rabbit hole, I was thinking about how to fix this, and I realized
that the simplest way is to add a marker on the OAuthException (OAuthException.oauth) and check for
it instead of using instance of. Resolves #82

https://github.com/jasonraimondi/ts-oauth2-server/issues/82#issuecomment-2250657230